+++ src/menus/link/create-panel-conf.ts
@@ -3,13 +3,13 @@
  * @author wangfupeng
  */
 
-import editor from '../../editor/index'
+import Editor from '../../editor/index'
 import { PanelConf } from '../menu-constructors/Panel'
 import { getRandom } from '../../utils/util'
 import $, { DomElement } from '../../utils/dom-core'
 import isActive from './is-active'
 
-export default function (editor: editor, text: string, link: string): PanelConf {
+export default function (editor: Editor, text: string, link: string): PanelConf {
     // panel 中需要用到的id
     const inputLinkId = getRandom('input-link')
     const inputTextId = getRandom('input-text')
@@ -79,7 +79,7 @@ export default function (editor: editor, text: string, link: string): PanelConf 
             return true
         } else {
             //用户未能通过开发者的校验，开发者希望我们提示这一字符串
-            alert(check)
+            editor.config.customAlert(check, 'warning')
         }
         return false
     }
@@ -95,18 +95,18 @@ export default function (editor: editor, text: string, link: string): PanelConf 
                 title: editor.i18next.t('menus.panelMenus.link.链接'),
                 // 模板
                 tpl: `<div>
-                        <input 
-                            id="${inputTextId}" 
-                            type="text" 
-                            class="block" 
-                            value="${text}" 
+                        <input
+                            id="${inputTextId}"
+                            type="text"
+                            class="block"
+                            value="${text}"
                             placeholder="${editor.i18next.t('menus.panelMenus.link.链接文字')}"/>
                         </td>
-                        <input 
-                            id="${inputLinkId}" 
-                            type="text" 
-                            class="block" 
-                            value="${link}" 
+                        <input
+                            id="${inputLinkId}"
+                            type="text"
+                            class="block"
+                            value="${link}"
                             placeholder="${editor.i18next.t('如')} https://..."/>
                         </td>
                         <div class="w-e-button-container">