+++ src/menus/quote/index.ts
@@ -44,6 +44,8 @@ class Quote extends BtnMenu implements MenuActive {
             })
             $topNodeElem.remove()
             editor.selection.moveCursor($targetELem.elems[len - 1])
+            // 即时更新btn状态
+            this.tryChangeActive()
         } else {
             // 将 P 转换为 quote
             const $quote = createQuote(topNodeElem)
@@ -54,6 +56,8 @@ class Quote extends BtnMenu implements MenuActive {
             moveNode.textContent
                 ? editor.selection.moveCursor(moveNode)
                 : editor.selection.moveCursor(moveNode, true)
+            // 即时更新btn状态
+            this.tryChangeActive()
             // 防止最后一行无法跳出
             $(`<p><br></p>`).insertAfter($quote)
             return
@@ -71,8 +75,8 @@ class Quote extends BtnMenu implements MenuActive {
      */
     public tryChangeActive(): void {
         const editor = this.editor
-        const cmdValue = editor.cmd.queryCommandValue('formatBlock')
-        if (cmdValue === 'blockquote') {
+        const cmdValue = editor.selection.getSelectionRangeTopNodes(editor)[0]?.getNodeName()
+        if (cmdValue === 'BLOCKQUOTE') {
             this.active()
         } else {
             this.unActive()
@@ -87,7 +91,7 @@ class Quote extends BtnMenu implements MenuActive {
     private getTopNodeName(): string {
         const editor = this.editor
         const $topNodeElem = editor.selection.getSelectionRangeTopNodes(editor)[0]
-        const nodeName = $topNodeElem.getNodeName()
+        const nodeName = $topNodeElem?.getNodeName()
 
         return nodeName
     }